% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcfboardr.R
\docType{package}
\name{gcfboardr}
\alias{gcfboardr}
\alias{gcfboardr-package}
\title{gcfboardr: A text dataset harvested from Green Climate Fund Board documents.}
\description{
The package also includes the script used to harvest data from the pdfs.
}
